requirements.txt
google-adk
litellm
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

